% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_plot.R
\name{batch_plot}
\alias{batch_plot}
\title{batch_plot}
\usage{
batch_plot(df, qc.label = "QC", xl, yl, tit)
}
\arguments{
\item{df}{dataframe containing the data}
}
\description{
plot sample values along batch
Plotting function
}
