% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_sdc.R
\name{pseudo_sdc}
\alias{pseudo_sdc}
\title{pseudo_sdc}
\usage{
pseudo_sdc(
  df = NULL,
  seed = 123,
  n.cores = 1,
  train.batch = NULL,
  test.breaks = NULL,
  test.window = NULL,
  test.index = NULL,
  criteria = "RSD",
  qc.label = NULL,
  qc.multibatch = FALSE,
  min.qc = NULL,
  quantile_increment = 0.1,
  log_transform = TRUE
)
}
\arguments{
\item{df}{The dataframe containing peak data. At minimum should contain columns labeled: name, sample, batch, compound, area, experiment_index, batch_index. \cr
Accepts only one compound at a time. If there are multiple compounds are present, run function in a loop, or lapply.}

\item{seed}{\code{numeric()} The seed to be set for reproducible results}

\item{n.cores}{\code{numeric()} The number of cores to be used for processing if being run on a multi-core machine}

\item{train.batch}{\code{character()} The batch name in the df which contains QC samples (should only be one batch which will be used for training regression spline model).}

\item{test.breaks}{\code{numeric()} A numeric vector indicating the number of equal sized sub-batches for the train.batch to be divided into and tested.}

\item{test.window}{\code{numeric()} A numeric vector containing the sizes of sliding windows to test when performing the sliding window median calculation.}

\item{test.index}{\code{numeric()} A numeric vector containing the injection position offset for pseudo QC inclusion in the peak data matrix.}

\item{criteria}{\code{character()} What criteria should be minimized when determining the optimal set of parameters. Should be one of:
\itemize{
\item "RSD" relative standard deviation
\item "MSE" mean squared error
\item "TSS" total sum of squares
}}

\item{qc.label}{\code{character()} Label designating the QC sample in the sample column of df.}

\item{min.qc}{\code{numeric()} The minimum number of pseudo-QC samples to consider during model training. Should be a value greater than 2.}

\item{quantile_increment}{\code{numeric()} Incremental step for qunatiles of peak areas to retain in training model.}

\item{log_transform}{\code{logical()} TRUE(default)/FALSE should data be log transformed}
}
\value{
\code{list()} containing:
\itemize{
\item df (original input data)
\item df_pseudoQC (data with pseudoQC calculated samples included). Includes an additional column labeled 'class' which categorizes true QC, Sample, Pseudo_QC samples.
\item df_pseudoQC_corrected (signal drift corrected data using pseudoQC samples). Same columns as df_pseudoQC returned, with an aditional 'area_corrected' column designating the signal drift corrected data.
\item criteria_table (table with results for criteria applied along with the other two non-used).
}
}
\description{
Signal drift correction using QC samples present in some batches but absent in others.
}
\examples{
sdc_out = pseudo_sdc(df = tog,
                     train.batch = "Run 4",
                     test.breaks = seq(10,25,2),
                     test.window = seq(1,10,2),
                     test.index = seq(0,35,5),
                     qc.label = "Pool",
                     min.qc = 10)

list2env(sdc_out ,.GlobalEnv)
}
