% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_outlier.R
\name{pw_outlier}
\alias{pw_outlier}
\title{pw_outlier}
\usage{
pw_outlier(
  df = NULL,
  n.cores = 1,
  mad_threshold = 3,
  pw_threshold = 0.95,
  peak_shrinkage = TRUE,
  grouping_factor = "batch",
  return_plot = FALSE,
  plot_name = "pw_outlier_plot",
  samps_exclude = "QC"
)
}
\arguments{
\item{df}{The dataframe containing peak data. At minimum it should contain columns labeled: name, sample, batch, compound, area, rep, rep_tech \cr
Additional columns are OK, but will not be used.}

\item{n.cores}{\code{numeric()} The number of cores to be used for processing if being run on a multi-core machine}

\item{mad_threshold}{\code{numeric()} The median absolute deviation (MAD) threshold to be used when identifying potential outliers}

\item{pw_threshold}{\code{numeric()} The pairwise outlier difference threshold. Should be between 0 and 1. Default is set to 0.95, meaning 5\% of the data will be identified as a potential outlier.}

\item{peak_shrinkage}{\code{logical()} For samples surpassing the MAD threshold, should they be shrunk towards the distribution (TRUE) or completely removed from the analysis (FALSE). \cr
If set to TRUE (default), values are shrunk towards the distribution of samples and will maintain the same rankings of samples while reducing the overall distribution skew during the pairwise elimination calculation.}

\item{grouping_factor}{The column label containing the grouping factor from which pairwise differences will be calculated. Default is batch column.}

\item{return_plot}{\code{logical()} Should a density plot be returned showing the pairwise difference threshold distributions per compound. Default is FALSE.}

\item{plot_name}{\code{character()} If return_plot is TRUE, what is the name of the output .pdf file. This will be saved to the current working directory.}

\item{samps_exclude}{\code{character()} Label designating the QC sample in the sample column of df. These will not be subjected to pairwise outlier detection since they can be useful downstream, for example with signal drift correction.}
}
\value{
\code{list()} containing:
\itemize{
\item df (original input data)
\item df_cleaned (pairwise outlier cleaned data).
\item df_rm (samples removed by the pairwise outlier elimination).
\item batch_plots plots of quantile threshold for each batch. Only returned if return_plot is set to TRUE.
}
}
\description{
Pairwise outlier removal of replicate samples within analytical batches. \cr
This is useful for example to identify technical errors, particularly when there is not extensive replication among samples to conduct more conventional outlier detection.
}
\examples{
pw_out = pw_outlier(
df = dat,
samps_exclude = "QC",
n.cores = 2,
mad_threshold = 3,
pw_threshold = 0.95,
peak_shrinkage = TRUE,
grouping_factor = "batch",
return_plot = FALSE,
plot_name = "pw_outlier_plot")

list2env(pw_out ,.GlobalEnv)
}
