% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf2Index.R
\name{sdf2index}
\alias{sdf2index}
\title{sdf2index}
\usage{
sdf2index(sdf_input, out_name)
}
\arguments{
\item{sdf_input}{Input file in .sdf format.}

\item{out_name}{Name of the .xls file to be written out.}
}
\value{
Two files written to the current working directory: 1) an index .xls file with the \code{outname} prefix, 2) an sdf file with the \code{valid} prefix.
}
\description{
Ensure data is in valid sdf format then index the sdf file. sdf files can be downloaded from https://mona.fiehnlab.ucdavis.edu/downloads. \cr
This function makes use of \link[ChemmineR]{} functions to ensure validity and formatting is correct for pseudoDrift
}
\examples{
sdf2index(sdf_input = system.file("extdata", "test.sdf", package = "pseudoDrift"), out_name = "Index.xls")
}
