% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  db_ids = NULL,
  compound_names = NULL,
  xls_file_name = NULL,
  valid_sdf_file = NULL,
  nbatch = 3,
  nsamps_per_batch = c(100, 50, 200),
  QC_freq = c(20, 10, 50),
  multiplyer = 100,
  seed = 123,
  sim_sd = NULL,
  m_eff = 1.5,
  b_eff = 3,
  save_rds = FALSE,
  rds_name = NULL
)
}
\arguments{
\item{db_ids}{\code{character()} ID(s) of compounds to be read from a validated sdf file.}

\item{compound_names}{\code{character()} Name of the compound(s) to be queried. Not case sensitive, but will only search compounds which begin with the string entered.}

\item{xls_file_name}{\code{character(1)} Index file in .xls format. \cr
This file can be produced using \link[pseudoDrift]{sdf2Index}.}

\item{valid_sdf_file}{\code{character(1)} sdf file to be read. \cr
This file can also be produced using \link[pseudoDrift]{sdf2Index}.}

\item{nbatch}{\code{numeric()} Number of batches to simulate.}

\item{nsamps_per_batch}{\code{numeric()} A numeric vector the same length as 1:nbatch. Number of samples included in each batch to be simulated.}

\item{QC_freq}{\code{numeric()} A numeric vector the same length as 1:nbatch. Frequency of QC samples.}

\item{multiplyer}{\code{numeric(1)} value to multiply exact m/z value from MoNA database to have an area measure (default is 1e2)}

\item{seed}{\code{numeric()} the seed to be used for reproducibility..}

\item{sim_sd}{\code{numeric()} Standard deviation of simulated data. Default value is 0.25*m/z value after multiplier.}

\item{m_eff}{\code{numeric()} maximum monotonic effect (slope).}

\item{b_eff}{\code{numeric()} maximum batch effect (between batch mean differences)}

\item{save_rds}{\code{logical(1)} To write to an rds file.}

\item{rds_name}{\code{character(1)} Name of the .rds file if save_rds is set to TRUE (default is FALSE).}
}
\value{
A nested tibble. The first column is the db id of the compounds simulated. The second column is the simulated matrix. The following four columns represent four batch simulation scenarios, which are: \cr
\itemize{
\item 1) t1_sim_mat - a monotonic (up/down effect) by batch.
\item 2) t2_sim_mat - a batch to batch block effect.
\item 3) t3_sim_mat - random, no systematic change.
\item 4) t4_sim_mat - monotonic and batch to batch block effect.
}
}
\description{
Simulate LC-MS/MS data or any other spectra from sdf db files.
}
\examples{
sim1 = simulate_data(compound_names = "tricin",
                     xls_file_name = system.file("extdata", "Index.xls", package = "pseudoDrift"),
                     valid_sdf_file = system.file("extdata", "valid-test.sdf", package = "pseudoDrift"))

sim2 = simulate_data(db_ids = c("FIO00738", "FIO00739","FIO00740"),
                     xls_file_name = system.file("extdata", "Index.xls", package = "pseudoDrift"),
                     valid_sdf_file = system.file("extdata", "valid-test.sdf", package = "pseudoDrift"))

## Providing name or compound db IDs gives the same result
identical(sim1,sim2)
}
